Test the validator systematically, and stop it panicking on a nil term
The validator's tests were spread across expression_test.go and fields_test.go,
so nothing led from validate.go to them and the file looked untested. They are now
in validate_test.go, and the tables they contained are joined by the systematic
half.

The tables enumerate cases someone thought of. These walk the vocabulary itself:
every declared operator must have a case in validateCall rather than being
reported to callers as unknown, every level and value type must be accepted, and
every field the enumeration offers must validate at its own level. An operator or
level added without a matching case now fails here instead of reaching a caller.
A second group buries one bad reference along each recursive path — under and, or
and not, in either operand of a comparison, as the subject of in, inside a nested
call and inside a some predicate — because a validator that only checks the root
is worth little.

The last group asserts only that validation returns, for trees a caller should not
be able to build but a decoder or an interceptor rewriting a filter can: a nil
argument, a nil term, a list where a predicate belongs. It found a crash on the
first run. A typed-nil *Call inside an `and` satisfies the type assertion and was
then dereferenced, and the same held for a nil reference, constant and list, so
each is now checked where its pointer is first read. Validation is the thing that
is supposed to answer for a malformed tree, so it must not fall over on one.
